requirements.txt
asgiref==3.2.3
Django==3.0.2
django-ckeditor==5.9.0
django-jalali-date==0.3.0
django-js-asset==1.2.2
django-social-share==1.3.2
jdatetime==3.6.2
Pillow==7.0.0
pytz==2019.3
sorl-thumbnail==12.6.2
sqlparse==0.3.0
